refactor: move html exporter unified cli
